+++ heksher/db_logic/rule.py
@@ -1,18 +1,16 @@
 from __future__ import annotations
 
-from asyncio.tasks import gather
 from datetime import datetime
 from itertools import groupby
 from operator import itemgetter
 from typing import Any, Dict, List, NamedTuple, Optional, Sequence, Tuple
 
 import orjson
-from sqlalchemy import join, select
+from sqlalchemy import Integer, and_, cast, func, join, not_, select, tuple_
+from sqlalchemy.orm import aliased
 
 from heksher.db_logic.logic_base import DBLogicBase
 from heksher.db_logic.metadata import conditions, context_features, rules, settings
-# This class should be used in the context of db logic api only
-from heksher.db_logic.util import inline_sql
 
 
 class RuleSpec(NamedTuple):
@@ -41,18 +39,19 @@ class RuleMixin(DBLogicBase):
         Returns:
             A RuleSpec describing the rule with the id, or None if no such rule exists.
         """
-        basic_results, feature_values = await gather(
-            self.db.fetch_one(
-                select([rules.c.setting, rules.c.value, rules.c.metadata]).where(rules.c.id == id_)
-            ),
-            self.db.fetch_all(
+        async with self.db_engine.connect() as conn:
+            basic_results = (await conn.execute(
+                select([rules.c.setting, rules.c.value, rules.c.metadata]).where(rules.c.id == id_).limit(1)
+            )).mappings().first()
+            feature_values = (await conn.execute(
                 select([conditions.c.context_feature, conditions.c.feature_value])
                 .select_from(join(conditions, context_features,
                                   conditions.c.context_feature == context_features.c.name))
                 .where(conditions.c.rule == id_)
                 .order_by(context_features.c.index)
-            )
-        )
+                .limit(1))
+                              ).mappings().all()
+
         # we query both the rule and its feature values at the same time. Meaning that if the rule does not exist,
         # we make 1 too many calls. However, we expect to make so few get_rule calls to non-existent rules that this
         # is negligible
@@ -80,29 +79,25 @@ class RuleMixin(DBLogicBase):
 
         """
         condition_count = len(match_conditions)
-        # we convert the conditions to a format that they can be queried with an "IN" condition
-        # (no, parameterization doesn't work)
-        condition_tuples = ','.join(
-            f"({inline_sql(k)},{inline_sql(v)})" for (k, v) in match_conditions.items()
-        )
-        query_template = f'''
-        SELECT DISTINCT id
-        FROM (
-          SELECT * from rules
-          WHERE setting = :setting
-        ) as rules
-        WHERE (SELECT COUNT(*) FROM conditions WHERE rule = rules.id) = :condition_count
-        AND NOT EXISTS (
-          SELECT *
-          from conditions
-          WHERE rule = id AND (context_feature, feature_value) NOT IN ({condition_tuples})
-        );
-        '''
+        condition_tuples = [(k, v) for (k, v) in match_conditions.items()]
 
-        return await self.db.fetch_val(
-            query_template,
-            {'setting': setting, 'condition_count': condition_count}
-        )
+        setting_rules = rules.select().where(rules.c.setting == setting)
+
+        async with self.db_engine.connect() as conn:
+            stmt = select(setting_rules.c.id.distinct()) \
+                .where(
+                and_(
+                    cast(select(func.count()).select_from(conditions)
+                         .where(conditions.c.rule == setting_rules.c.id), Integer) == condition_count,
+                    not_(conditions.select()
+                         .where(and_(conditions.c.rule == setting_rules.c.id,
+                                     tuple_(conditions.c.context_feature, conditions.c.feature_value)
+                                     .not_in(condition_tuples)))
+                         .exists())
+                )
+            )
+            resp = await conn.execute(stmt)
+        return resp.scalar_one_or_none()
 
     async def delete_rule(self, rule_id: int):
         """
@@ -110,7 +105,8 @@ class RuleMixin(DBLogicBase):
         Args:
             rule_id: the id of the rule to delete
         """
-        await self.db.execute(rules.delete().where(rules.c.id == rule_id))
+        async with self.db_engine.begin() as conn:
+            await conn.execute(rules.delete().where(rules.c.id == rule_id))
 
     async def add_rule(self, setting: str, value: Any, metadata: Dict[str, Any],
                        match_conditions: Dict[str, str]) -> int:
@@ -130,15 +126,16 @@ class RuleMixin(DBLogicBase):
         """
         metadata_ = str(orjson.dumps(metadata), 'utf-8')
         value_ = str(orjson.dumps(value), 'utf-8')
-        async with self.db.transaction():
-            rule_id = await self.db.fetch_val(
+        async with self.db_engine.begin() as conn:
+            rule_id = (await conn.execute(
                 rules.insert()
                 .values(setting=setting, value=value_, metadata=metadata_)
                 .returning(rules.c.id)
-            )
-            await self.db.execute_many(
-                conditions.insert(),
-                [{'rule': rule_id, 'context_feature': k, 'feature_value': v} for (k, v) in match_conditions.items()]
+            )).scalar_one()
+            await conn.execute(
+                conditions.insert().values(
+                    [{'rule': rule_id, 'context_feature': k, 'feature_value': v}
+                     for (k, v) in match_conditions.items()])
             )
         return rule_id
 
@@ -150,7 +147,8 @@ class RuleMixin(DBLogicBase):
             value: Value to change to
         """
         encoded_value = str(orjson.dumps(value), 'utf-8')
-        await self.db.execute(rules.update().where(rules.c.id == rule_id).values(value=encoded_value))
+        async with self.db_engine.begin() as conn:
+            await conn.execute(rules.update().where(rules.c.id == rule_id).values(value=encoded_value))
 
     async def query_rules(self, setting_names: List[str],
                           feature_value_options: Optional[Dict[str, Optional[List[str]]]],
@@ -164,7 +162,7 @@ class RuleMixin(DBLogicBase):
             feature_value_options: The options for each context feature. Rules that cannot match with these options are
              discounted. If None, all rules are counted
             setting_touch_time_cutoff: If provided, will discount all rules pertaining to settings that have not been
-             updated since  this time.
+             updated since this time.
             include_metadata: Whether to retrieve and include the metadata of each rule in the result.
 
         Returns:
@@ -172,18 +170,20 @@ class RuleMixin(DBLogicBase):
 
         """
         if setting_touch_time_cutoff:
-            settings_results = await self.db.fetch_all(
-                select([settings.c.name])
-                .where(
-                    settings.c.name.in_(setting_names)
-                    & (settings.c.last_touch_time >= setting_touch_time_cutoff)
-                )
-            )
+            async with self.db_engine.connect() as conn:
+                settings_results = (await conn.execute(
+                    select([settings.c.name])
+                    .where(
+                        settings.c.name.in_(setting_names)
+                        & (settings.c.last_touch_time >= setting_touch_time_cutoff)
+                    )
+                )).mappings().all()
             applicable_settings = [row['name'] for row in settings_results]
         else:
             applicable_settings = setting_names
 
         applicable_rules: Dict[int, Tuple[Tuple[str, str], ...]] = {}
+        conditions_ = aliased(conditions)
 
         if not applicable_settings:
             # shortcut in case all settings are up-to-date
@@ -193,49 +193,48 @@ class RuleMixin(DBLogicBase):
             # it will not be returned
             if feature_value_options is None:
                 # match all
-                inv_match = 'FALSE'
+                inv_match = False
             elif not feature_value_options:
                 # match none
-                inv_match = 'TRUE'
+                inv_match = True
             else:
                 exact_tuple_conditions = []
                 only_cf_conditions = []
                 for k, v in feature_value_options.items():
-                    # (no, parameterization doesn't work)
                     if v is None:
-                        only_cf_conditions.append(inline_sql(k))
+                        only_cf_conditions.append(k)
                     else:
                         for cf_value in v:
-                            exact_tuple_conditions.append(f"({inline_sql(k)},{inline_sql(cf_value)})")
-                inv_match_predicates = []
-                if exact_tuple_conditions:
-                    inv_match_predicates.append(
-                        '(context_feature, feature_value) NOT IN (' + ','.join(exact_tuple_conditions) + ')'
-                    )
-                if only_cf_conditions:
-                    inv_match_predicates.append('context_feature NOT IN (' + ','.join(only_cf_conditions) + ')')
-                inv_match = ' AND '.join(inv_match_predicates)
+                            exact_tuple_conditions.append((k, cf_value))
+                tuple_conditions = tuple_(conditions_.c.context_feature, conditions_.c.feature_value).not_in(
+                    exact_tuple_conditions)
+                cf_conditions = conditions_.c.context_feature.not_in(only_cf_conditions)
+                if exact_tuple_conditions and only_cf_conditions:
+                    inv_match = and_(tuple_conditions, cf_conditions)
+                else:
+                    inv_match = tuple_conditions if exact_tuple_conditions else cf_conditions
 
-            # (no, parameterization doesn't work)
-            settings_container = ','.join(f"'{name}'" for name in applicable_settings)
-            query = f"""
-            SELECT rules.id, C.context_feature, C.feature_value -- get all the conditions for all the rules
-            FROM
-            (
-                (SELECT * from rules where setting IN ({settings_container})) as rules
-                LEFT JOIN
-                conditions as C
-                ON rules.id = C.rule
-            )
-            LEFT JOIN context_features ON context_features.name = C.context_feature
-            WHERE NOT EXISTS (
-              SELECT *
-              from conditions
-              WHERE rule = C.rule AND {inv_match}
-            )
-            ORDER BY rules.id, context_features.index;
-            """
-            conditions_results = await self.db.fetch_all(query)
+            settings_container = [name for name in applicable_settings]
+
+            query = select() \
+                .add_columns(rules.c.id, conditions.c.context_feature, conditions.c.feature_value) \
+                .select_from(rules.outerjoin(conditions, rules.c.id == conditions.c.rule)) \
+                .outerjoin(context_features, context_features.c.name == conditions.c.context_feature) \
+                .where(
+                and_(rules.c.setting.in_(settings_container),
+                     not_(conditions_.select()
+                          .where(
+                         and_(
+                             conditions.c.rule == conditions_.c.rule, inv_match
+                         ))
+                          .exists())
+                     )
+            ) \
+                .order_by(rules.c.id, context_features.c.index)
+
+            async with self.db_engine.connect() as conn:
+                conditions_results = (await conn.execute(query)
+                                      ).mappings().all()
             # group all the conditions by rules
             for rule_id, rows in groupby(conditions_results, key=itemgetter('id')):
                 rule_conditions = tuple((row['context_feature'], row['feature_value']) for row in rows)
@@ -254,7 +253,8 @@ class RuleMixin(DBLogicBase):
             )
             if include_metadata:
                 rule_query.append_column(rules.c.metadata)
-            rule_results = await self.db.fetch_all(rule_query)
+            async with self.db_engine.connect() as conn:
+                rule_results = (await conn.execute(rule_query)).mappings().all()
 
         ret = {}
         missing_settings = set(applicable_settings)