+++ tests/unittests/test_util.py
@@ -1,6 +1,6 @@
-from pytest import mark, raises
+from pytest import mark
 
-from heksher.db_logic.util import inline_sql, supersequence_new_elements
+from heksher.db_logic.util import supersequence_new_elements
 
 
 @mark.parametrize('a,b,new', [
@@ -27,24 +27,3 @@ def test_issupersequence(a, b, new):
 def test_is_not_supersequence(a, b):
     ss = supersequence_new_elements(a, b)
     assert ss is None
-
-
-@mark.parametrize('x', [
-    "a'",
-    "' OR '1' = '1",
-    "-- DROP TABLE x",
-    ''
-])
-def test_invalid_inline_sql(x):
-    with raises(AssertionError):
-        inline_sql(x)
-
-
-@mark.parametrize('x', [
-    "a",
-    "abra cadabra",
-    'DROP TABLE x',
-    "DROP TABLE x at 9 oclock"
-])
-def test_valid_inline_sql(x):
-    assert inline_sql(x) == "'" + x + "'"